+++ Documentation/reference/Danger.swift
@@ -3,6 +3,7 @@ import Darwin.C
 import Foundation
 import Logger
 import OctoKit
+import ShellOut
 import SwiftOnoneSupport
 
 public struct BitBucketServer: Decodable, Equatable {
@@ -27,7 +28,7 @@ public struct BitBucketServerActivity: Decodable, Equatable {
     public let id: Int
 
     /// Date activity created as number of mili seconds since the unix epoch
-    public let createdDate: Int
+    public let createdAt: Int
 
     /// The user that triggered the activity.
     public let user: Danger.BitBucketServerUser
@@ -44,7 +45,7 @@ public struct BitBucketServerComment: Decodable, Equatable {
     public let id: Int
 
     /// Date comment created as number of mili seconds since the unix epoch
-    public let createdDate: Int
+    public let createdAt: Int
 
     /// The comment's author
     public let user: Danger.BitBucketServerUser
@@ -68,9 +69,9 @@ public struct BitBucketServerComment: Decodable, Equatable {
     public let commentAction: String?
 
     /// Detailed data of the comment
-    public let comment: Danger.BitBucketServerComment.BitBucketServerCommentInner?
+    public let comment: Danger.BitBucketServerComment.CommentDetail?
 
-    public struct BitBucketServerCommentInner: Decodable, Equatable {
+    public struct CommentDetail: Decodable, Equatable {
         /// The comment's id
         public let id: Int
 
@@ -90,20 +91,20 @@ public struct BitBucketServerComment: Decodable, Equatable {
         public let updatedAt: Int
 
         /// Replys to the comment
-        public let comments: [Danger.BitBucketServerComment.BitBucketServerCommentInner]
+        public let comments: [Danger.BitBucketServerComment.CommentDetail]
 
         /// Properties associated with the comment
-        public let properties: Danger.BitBucketServerComment.BitBucketServerCommentInner.BitBucketServerCommentInnerProperties
+        public let properties: Danger.BitBucketServerComment.CommentDetail.InnerProperties
 
         /// Tasks associated with the comment
-        public let tasks: [Danger.BitBucketServerComment.BitBucketServerCommentInner.BitBucketServerCommentTask]
+        public let tasks: [Danger.BitBucketServerComment.CommentDetail.BitBucketServerCommentTask]
 
         public struct BitBucketServerCommentTask: Decodable, Equatable {
             /// The tasks ID
             public let id: Int
 
             /// Date activity created as number of mili seconds since the unix epoch
-            public let createdDate: Int
+            public let createdAt: Int
 
             /// The text of the task
             public let text: String
@@ -115,7 +116,7 @@ public struct BitBucketServerComment: Decodable, Equatable {
             public let author: Danger.BitBucketServerUser
         }
 
-        public struct BitBucketServerCommentInnerProperties: Decodable, Equatable {
+        public struct InnerProperties: Decodable, Equatable {
             /// The ID of the repo
             public let repositoryId: Int
 
@@ -175,10 +176,10 @@ public struct BitBucketServerMergeRef: Decodable, Equatable {
 
 public struct BitBucketServerMetadata: Decodable, Equatable {
     /// The PR's ID
-    public let pullRequestID: String
+    public var pullRequestID: String { get }
 
     /// The complete repo slug including project slug.
-    public let repoSlug: String
+    public var repoSlug: String { get }
 }
 
 public struct BitBucketServerPR: Decodable, Equatable {
@@ -212,7 +213,7 @@ public struct BitBucketServerPR: Decodable, Equatable {
     /// The PR submittor's reference
     public let fromRef: Danger.BitBucketServerMergeRef
 
-    /// The repo Danger is sunning on
+    /// The repo Danger is running on
     public let toRef: Danger.BitBucketServerMergeRef
 
     /// Is the PR locked?
@@ -225,7 +226,7 @@ public struct BitBucketServerPR: Decodable, Equatable {
     public let reviewers: [Danger.BitBucketServerUser]
 
     /// People who have participated in the PR
-    public let participants: [Danger.BitBucketServerUser]
+    public let participants: [Danger.BitBucketServerPR.BitBucketServerAuthor]
 
     public struct BitBucketServerAuthor: Decodable, Equatable {
         /// The BitBucket Server User
@@ -304,7 +305,7 @@ public struct DangerDSL: Decodable {
 
     public let github: Danger.GitHub!
 
-    public let bitbucket_server: Danger.BitBucketServer!
+    public let bitbucketServer: Danger.BitBucketServer!
 
     public let utils: Danger.DangerUtils
 
@@ -370,7 +371,7 @@ extension DangerDSL {
 public struct DangerUtils {
     /// Let's you go from a file path to the contents of the file
     /// with less hassle.
-    /// Tt specifically assumes golden path code so Dangerfiles
+    /// It specifically assumes golden path code so Dangerfiles
     /// don't have to include error handlings - an error will
     /// exit evaluation entirely as it should only happen at dev-time.
     ///
@@ -407,6 +408,8 @@ public enum FileType: String, Equatable {
 
     case yml
 }
+
+extension FileType: CaseIterable {}
 
 extension FileType {
     public var `extension`: String { get }
@@ -699,6 +702,9 @@ public struct GitHubPR: Decodable, Equatable {
 
     /// The milestone of the pull request
     public let milestone: Danger.GitHubMilestone?
+
+    /// The link back to this PR as user-facing
+    public let htmlUrl: String
 }
 
 public struct GitHubRepo: Decodable, Equatable {
@@ -790,6 +796,29 @@ public struct GitHubUser: Decodable, Equatable {
 
     /// The type of user: user or organization.
     public let userType: Danger.GitHubUser.UserType
+}
+
+/// Meta information for showing in the text info
+public struct Meta: Codable {}
+
+/// The SwiftLint plugin has been embedded inside Danger, making
+/// it usable out of the box.
+public struct SwiftLint {
+    /// This is the main entry point for linting Swift in PRs using Danger-Swift.
+    /// Call this function anywhere from within your Dangerfile.swift.
+    public static func lint(inline: Bool = default, directory: String? = default, configFile: String? = default, lintAllFiles: Bool = default, swiftlintPath: String = default) -> [Danger.SwiftLintViolation]
+}
+
+public struct SwiftLintViolation: Codable {
+    /// Creates a new instance by decoding from the given decoder.
+    ///
+    /// This initializer throws an error if reading from the decoder fails, or
+    /// if the data read is corrupted or otherwise invalid.
+    ///
+    /// - Parameter decoder: The decoder to read data from.
+    public init(from decoder: Decoder) throws
+
+    public func toMarkdown() -> String
 }
 
 /// The result of a warn, message, or fail.