+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## Master
 
+- Fix danger report system by [@f-meloni][]
 - More docs by [@orta][]
 - Fixes for people installing via homebrew (thanks [@f-meloni][]) by [@orta][]